transcoder/requirements.txt
boto3
pydantic_settings
python-dotenv
requests